Added customizable background color to Stars component
The Stars component now supports a new prop, 'bgColour', which allows the background color to be customized. This change provides more flexibility in using the Stars component across different contexts and themes. The default value is set as 'bg-neutral-950'. Updated usage of this component in +page.svelte file to reflect this new feature.