Updated image sources and added looping to video
In this update, we've replaced the external image sources with local ones for better control and performance. We also made a change to the video element on the page, adding a loop attribute so that it plays continuously. Additionally, some adjustments were made to CSS classes for improved layout and positioning of elements. New static files have been added as well.