Updated video and image styling
Adjusted the CSS classes for the video and image elements to improve layout. The z-index of the video element was increased, while its max-height class was removed. The Background component was commented out in SvelteFlow. For images, responsiveness was improved by making width full on smaller screens and 30rem on larger ones.