Updated presale date and enhanced banner UI
The presale date has been updated in the constants file. The banner component has been significantly improved with a countdown timer to the presale, alert badges, and links to join the whitelist. A minor change was made in the schema file to ensure newline at end of file. In addition, Blur SVG is now imported as both a component and source in the presale page. Lastly, an additional color 'pill' was added to brand colors in Tailwind configuration.